Module Irmin_IDB

An Irmin backend that stores data in an IndexedDB.

val config : string -> Irmin.config

module AO : functor (K : Irmin.Hash.S) -> Irmin_git.VALUE_STORE

An append-only store that stores Git.Value.t items using hash K.

exception Format_too_old of [ `Irmin_0_10 ]

Raised on Repo.create if the existing data format cannot be read by this version of irmin-indexeddb. To migrate Irmin 0.10 format data, upgrade to irmin-indexeddb version 0.6 first.

module Make : functor (C : Irmin.Contents.S) -> functor (P : Irmin.Path.S) -> functor (B : Irmin.Branch.S) -> Irmin.S with type Make.key = P.t and type Make.step = P.step and type Make.metadata = Irmin_git.Metadata.t and type Make.contents = C.t and type Make.branch = B.t and type Make.Commit.Hash.t = Irmin.Hash.SHA1.t and type Make.Tree.Hash.t = Irmin.Hash.SHA1.t and type Make.Contents.Hash.t = Irmin.Hash.SHA1.t